refactor(hub): move with-shape/i18n → shape/via-i18n (#623)
