+++ packages/hub/__tests__/attestation-vault.test.ts
@@ -1,10 +1,10 @@
 import { describe, it, expect } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
 import { verifyAttestation } from '@noy-db/attestation'
-import { withI18n } from '../src/with-shape/i18n/index.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
 import { withHistory } from '../src/with-commit/history/index.js'
 import { withAttestation } from '../src/with-audit/attestation/index.js'
-import { i18nText } from '../src/with-shape/i18n/core.js'
+import { i18nText } from '../src/shape/via-i18n/core.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ConflictError, AttestationError } from '../src/kernel/errors.js'
 import { withTeam } from '../src/with-party/team/index.js'
