+++ packages/hub/__tests__/describe-contract.test.ts
@@ -16,10 +16,10 @@
 import { describe, it, expect } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
 import { money } from '../src/shape/via-money/descriptor.js'
-import { staticDict } from '../src/with-shape/i18n/dictionary.js'
-import { i18nText } from '../src/with-shape/i18n/core.js'
+import { staticDict } from '../src/shape/via-i18n/dictionary.js'
+import { i18nText } from '../src/shape/via-i18n/core.js'
 import { ref } from '../src/kernel/refs.js'
-import { withI18n } from '../src/with-shape/i18n/active.js'
+import { withI18n } from '../src/shape/via-i18n/active.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import type { CollectionDescription } from '@noy-db/hub/describe'
 import { ConflictError } from '../src/kernel/errors.js'
