+++ packages/hub/__tests__/dictkey-parity.test.ts
@@ -4,8 +4,8 @@
  */
 import { describe, it, expect, beforeEach } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
-import { dictKey } from '../src/with-shape/i18n/dictionary.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
+import { dictKey } from '../src/shape/via-i18n/dictionary.js'
 import { ConflictError } from '../src/kernel/errors.js'
 import type { Noydb } from '../src/kernel/noydb.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
