+++ packages/hub/__tests__/i18n-layers.test.ts
@@ -16,8 +16,8 @@
  */
 import { describe, it, expect } from 'vitest'
 import { createNoydb, withGuard, withDerivation } from '../src/index.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
-import { i18nText } from '../src/with-shape/i18n/core.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
+import { i18nText } from '../src/shape/via-i18n/core.js'
 import type { NoydbStore, EncryptedEnvelope } from '../src/kernel/types.js'
 
 function memory(): NoydbStore {
