+++ packages/hub/__tests__/i18n-read-layer.test.ts
@@ -7,8 +7,8 @@
  */
 import { describe, it, expect, beforeEach } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
-import { i18nText } from '../src/with-shape/i18n/core.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
+import { i18nText } from '../src/shape/via-i18n/core.js'
 import { ConflictError } from '../src/kernel/errors.js'
 import type { Noydb } from '../src/kernel/noydb.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
