+++ packages/hub/__tests__/i18n.test.ts
@@ -15,7 +15,7 @@
 
 import { describe, it, expect, beforeEach } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
 import type { Noydb } from '../src/kernel/noydb.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ConflictError } from '../src/kernel/errors.js'
@@ -23,7 +23,7 @@ import {
   MissingTranslationError,
   LocaleNotSpecifiedError,
 } from '../src/kernel/errors.js'
-import { i18nText, applyI18nLocale, resolveI18nText, validateI18nTextValue } from '../src/with-shape/i18n/core.js'
+import { i18nText, applyI18nLocale, resolveI18nText, validateI18nTextValue } from '../src/shape/via-i18n/core.js'
 
 // ─── Inline memory adapter ─────────────────────────────────────────────
