+++ packages/hub/__tests__/introspection/json-schema.test.ts
@@ -11,7 +11,7 @@ import { describe, it, expect } from 'vitest'
 import { z } from 'zod'
 import { createNoydb } from '../../src/kernel/noydb.js'
 import { money } from '../../src/shape/via-money/descriptor.js'
-import { dictKey } from '../../src/with-shape/i18n/dictionary.js'
+import { dictKey } from '../../src/shape/via-i18n/dictionary.js'
 import { ref } from '../../src/kernel/refs.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../../src/kernel/types.js'
 import { ConflictError } from '../../src/kernel/errors.js'
