+++ packages/hub/__tests__/materialized-views/i18n.test.ts
@@ -11,8 +11,8 @@ import { describe, it, expect } from 'vitest'
 import { createNoydb, withMaterializedView } from '../../src/index.js'
 import { withAggregate } from '../../src/with-lookup/aggregate/index.js'
 import { sum, count } from '../../src/with-lookup/aggregate/reducers.js'
-import { i18nText } from '../../src/with-shape/i18n/core.js'
-import { withI18n } from '../../src/with-shape/i18n/index.js'
+import { i18nText } from '../../src/shape/via-i18n/core.js'
+import { withI18n } from '../../src/shape/via-i18n/index.js'
 import { LocaleNotSpecifiedError, MaterializedViewConfigError } from '../../src/kernel/errors.js'
 import type { NoydbStore, EncryptedEnvelope } from '../../src/kernel/types.js'
