+++ packages/hub/__tests__/query-dictkey.test.ts
@@ -10,9 +10,9 @@ import { describe, it, expect } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ConflictError } from '../src/kernel/errors.js'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
 import { withAggregate } from '../src/with-lookup/aggregate/index.js'
-import { dictKey } from '../src/with-shape/i18n/dictionary.js'
+import { dictKey } from '../src/shape/via-i18n/dictionary.js'
 import { sum, count } from '../src/kernel/query/index.js'
 
 function memory(): NoydbStore {
