+++ packages/hub/__tests__/query-join-i18n.test.ts
@@ -8,8 +8,8 @@ import { createNoydb } from '../src/kernel/noydb.js'
 import { ConflictError } from '../src/kernel/errors.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/kernel/types.js'
 import { ref } from '../src/kernel/refs.js'
-import { i18nText } from '../src/with-shape/i18n/core.js'
-import { withI18n } from '../src/with-shape/i18n/index.js'
+import { i18nText } from '../src/shape/via-i18n/core.js'
+import { withI18n } from '../src/shape/via-i18n/index.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
