+++ packages/hub/src/kernel/collection-config.ts
@@ -31,7 +31,7 @@ import type { LedgerStore } from '../with-commit/history/ledger/index.js'
 import type { CrdtMode } from '../with-commit/crdt/crdt.js'
 import { NO_CRDT, type CrdtStrategy } from '../with-commit/crdt/strategy.js'
 import { NO_HISTORY, type HistoryStrategy } from '../with-commit/history/strategy.js'
-import { NO_I18N, type I18nStrategy } from '../with-shape/i18n/strategy.js'
+import { NO_I18N, type I18nStrategy } from '../shape/via-i18n/strategy.js'
 import { NO_SYNC, type SyncStrategy } from '../with-party/team/sync-strategy.js'
 import { NO_BLOBS, type BlobStrategy } from '../with-shape/blobs/strategy.js'
 import { NO_AGGREGATE, type AggregateStrategy } from '../with-lookup/aggregate/strategy.js'
@@ -41,8 +41,8 @@ import type { ObjectProjection } from '../with-shape/blobs/object-projection.js'
 import type { BlobFieldsConfig } from '../with-shape/blobs/blob-compaction.js'
 import type { IndexStrategy } from '../with-lookup/indexing/strategy.js'
 import type { IndexDef } from '../with-lookup/indexing/eager-indexes.js'
-import type { I18nTextDescriptor } from '../with-shape/i18n/core.js'
-import type { DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../with-shape/i18n/dictionary.js'
+import type { I18nTextDescriptor } from '../shape/via-i18n/core.js'
+import type { DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../shape/via-i18n/dictionary.js'
 import type { ComputedFields } from '../with-formula/computed/index.js'
 import { resolveClassifiedFields, ClassifiedConfigError, type ClassifiedEntry, type ResolvedClassified } from '../with-shape/classified/resolve.js'
 import { guardClassifiedCompat, type ClassifiedGuardCtx } from '../with-shape/classified/guards.js'
