+++ packages/hub/src/kernel/types.ts
@@ -49,8 +49,8 @@ import type { TeamStrategy } from '../port/with/team-strategy.js'
 import type { LazyStrategy } from '../port/with/lazy-strategy.js'
 import type { SearchStrategy } from '../with-lookup/search/strategy.js'
 import type { CargoStrategy } from '../with-cargo/strategy.js'
-import type { Layer } from '../with-shape/i18n/policy.js'
-import type { I18nStrategy } from '../with-shape/i18n/strategy.js'
+import type { Layer } from '../shape/via-i18n/policy.js'
+import type { I18nStrategy } from '../shape/via-i18n/strategy.js'
 import type { SessionStrategy } from '../with-party/session/strategy.js'
 import type { SyncStrategy } from '../with-party/team/sync-strategy.js'
 import type { GuardStrategyHandleAny } from '../with-audit/guards/types.js'
@@ -64,7 +64,7 @@ import type { SealingKeyProvider, RecipientHint } from '../with-party/team/manag
 import type { ShamirRecoveryProvider } from '../with-party/team/shamir-recovery-provider.js'
 import type { ObjectProjection } from '../with-shape/blobs/object-projection.js'
 import type { CoordinationProvider } from '../port/by/types.js'
-import type { ScriptWarning } from '../with-shape/i18n/script.js'
+import type { ScriptWarning } from '../shape/via-i18n/script.js'
 import type { ViaDescriptor } from './via.js'
 import type { EnclaveKey } from './enclave/index.js'
