+++ packages/hub/src/shape/via-i18n/densify.ts
