+++ packages/hub/src/shape/via-i18n/dictionary.ts
