+++ packages/hub/src/shape/via-i18n/script.ts
