+++ packages/hub/src/with-formula/materialized-views/types.ts
@@ -4,7 +4,7 @@ import type { AggregateSpec, Aggregation } from '../../with-lookup/aggregate/agg
 import type { GroupedAggregation } from '../../with-lookup/aggregate/groupby.js'
 import type { JoinStrategy } from '../../kernel/query/join.js'
 import type { MoneyDescriptor } from '../../shape/via-money/descriptor.js'
-import type { I18nTextDescriptor } from '../../with-shape/i18n/core.js'
+import type { I18nTextDescriptor } from '../../shape/via-i18n/core.js'
 
 /**
  * Minimal vault-shaped accessor passed to the MV `query()` callback.
