+++ packages/hub/src/with-shape/introspection/describe.ts
@@ -16,9 +16,9 @@ import { resolveFieldMeta, validateFieldMetaKeys, FieldMetaUnknownFieldError, hu
 import type { CollectionMeta } from './meta.js'
 import type { MoneyDescriptor } from '../../shape/via-money/descriptor.js'
 import type { ViaDescriptor } from '../../kernel/via.js'
-import type { DictKeyDescriptor, StaticDictDescriptor } from '../i18n/dictionary.js'
-import { isStaticDictDescriptor } from '../i18n/dictionary.js'
-import type { I18nTextDescriptor } from '../i18n/core.js'
+import type { DictKeyDescriptor, StaticDictDescriptor } from '../../shape/via-i18n/dictionary.js'
+import { isStaticDictDescriptor } from '../../shape/via-i18n/dictionary.js'
+import type { I18nTextDescriptor } from '../../shape/via-i18n/core.js'
 import type { ComputedFields } from '../../with-formula/computed/index.js'
 import type { RefDescriptor } from '../../kernel/refs.js'
 import type { ClassifiedFieldSpec } from '../classified/descriptor.js'
